LPD-91213 Inline, now rename scheduledPatchSEOStudioScans and _patchSEOStudioScan when we have more
